fixes cublasHandle_t problem, better softmax

+++ moses/FF/NMT/mblas/thrust_functions.h
@@ -15,7 +15,7 @@ namespace thrust
       template<typename T>
       struct unary_exp : public thrust::unary_function<T,T> {
         __host__ __device__
-        T operator()(const T &x) const { return exp(x); }
+        T operator()(const T &x) const { return expf(x); }
       };
       
       template<typename Eval>
@@ -28,7 +28,7 @@ namespace thrust
       template<typename T>
       struct unary_log : public thrust::unary_function<T,T> {
         __host__ __device__
-        T operator()(const T &x) const { return log(x); }
+        T operator()(const T &x) const { return logf(x); }
       };
       
       template<typename Eval>
@@ -41,7 +41,7 @@ namespace thrust
       template<typename T>
       struct unary_logit : public thrust::unary_function<T,T> {
         __host__ __device__
-        T operator()(const T &x) const { return 1.0 / (1.0 + exp(-x)); }
+        T operator()(const T &x) const { return 1.0 / (1.0 + expf(-x)); }
       };
       
       template<typename Eval>
@@ -54,7 +54,7 @@ namespace thrust
       template<typename T>
       struct unary_tanh : public thrust::unary_function<T,T> {
         __host__ __device__
-        T operator()(const T &x) const { return tanh(x); }
+        T operator()(const T &x) const { return tanhf(x); }
       };
       
       template<typename Eval>
